Fix issue with fetching the wrong page when updating filters
We need to reset the page to 1 when the filters change, otherwise we
fetch the wrong page and it might result in an error. Resetting the page
number only in the useEffect hook is not enough, because the filters in
state are updated before the useEffect and the query will use the page
number from the "old" useState.

Use a ref to store the previous filters value and determine if they have
changed. Use 1 as the value for the page number whenever the filters
change.