Change the way event details are navigated to
Remove react-router-dom since we are creating an embeddable widget and
we can't use hash routes in the app.

Add new data-attribute to decide if the event list or event detail page
should be shown.

This change requires two app deployments in two different URLs to be
active at the same time. The event list page can be deployed to the root
path and the event detail page will be deployed to the /event/ path.

For the event detail page, we parse the event ID from the URL and pass
it to the event query.